Refactor ManagedLedgerAttributes
