+++ managed-ledger/src/main/java/org/apache/bookkeeper/mledger/ManagedLedgerAttributes.java
@@ -16,32 +16,16 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.bookkeeper.mledger.impl;
+package org.apache.bookkeeper.mledger;
 
-import com.google.common.annotations.VisibleForTesting;
-import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.common.Attributes;
 import lombok.Data;
-import org.apache.bookkeeper.mledger.ManagedLedger;
 import org.apache.pulsar.common.naming.TopicName;
 import org.apache.pulsar.opentelemetry.OpenTelemetryAttributes;
+import org.apache.pulsar.opentelemetry.OpenTelemetryAttributes.ManagedLedgerOperationStatus;
 
 @Data
 public class ManagedLedgerAttributes {
-
-    public static final AttributeKey<String> PULSAR_MANAGER_LEDGER_NAME =
-            AttributeKey.stringKey("pulsar.managed_ledger.name");
-
-    public static final AttributeKey<String> PULSAR_MANAGED_LEDGER_OPERATION_STATUS =
-            AttributeKey.stringKey("pulsar.managed_ledger.operation.status");
-
-    @VisibleForTesting
-    public enum OperationStatus {
-        SUCCESS,
-        FAILURE;
-        public final Attributes attributes =
-                Attributes.of(PULSAR_MANAGED_LEDGER_OPERATION_STATUS, name().toLowerCase());
-    };
 
     private final Attributes attributes;
     private final Attributes attributesOperationSucceed;
@@ -51,16 +35,16 @@ public class ManagedLedgerAttributes {
         var mlName = ml.getName();
         var topicName = TopicName.get(TopicName.fromPersistenceNamingEncoding(mlName));
         attributes = Attributes.of(
-                PULSAR_MANAGER_LEDGER_NAME, ml.getName(),
+                OpenTelemetryAttributes.ML_NAME, mlName,
                 OpenTelemetryAttributes.PULSAR_NAMESPACE, topicName.getNamespace()
         );
         attributesOperationSucceed = Attributes.builder()
                 .putAll(attributes)
-                .putAll(OperationStatus.SUCCESS.attributes)
+                .putAll(ManagedLedgerOperationStatus.SUCCESS.attributes)
                 .build();
         attributesOperationFailure = Attributes.builder()
                 .putAll(attributes)
-                .putAll(OperationStatus.FAILURE.attributes)
+                .putAll(ManagedLedgerOperationStatus.FAILURE.attributes)
                 .build();
     }
 }
